Subversion: on Ubuntu 11.x we doesn't recieve full error text: Auth realm..\npassword or Auth realm\ncertificate:
So command will always be interrupted after 1 second waiting if error stream becomes not empty

IDEA-102670 Subversion: command line commit fails on attempt to commit to 2 repositories at the same time
(cherry picked from commit 40166fd)

+++ plugins/svn4idea/bindSvn/src/org/jetbrains/idea/svn/commandLine/SvnCommand.java
@@ -118,15 +118,21 @@ public abstract class SvnCommand {
 
   /**
    * Wait for process termination
+   * @param timeout
    */
-  public void waitFor() {
+  public boolean waitFor(int timeout) {
     checkStarted();
     final OSProcessHandler handler;
     synchronized (myLock) {
-      if (myIsDestroyed) return;
+      if (myIsDestroyed) return true;
       handler = myHandler;
     }
-    handler.waitFor();
+    if (timeout == -1) {
+      return handler.waitFor();
+    }
+    else {
+      return handler.waitFor(timeout);
+    }
   }
 
   protected abstract void processTerminated(int exitCode);
@@ -173,8 +179,10 @@ public abstract class SvnCommand {
 
   public void destroyProcess() {
     synchronized (myLock) {
-      myIsDestroyed = true;
-      myHandler.destroyProcess();
+      if (! myIsDestroyed) {
+        myIsDestroyed = true;
+        myHandler.destroyProcess();
+      }
     }
   }
